requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.10.0
astunparse==1.6.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
black==25.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
Chroma==0.2.0
chromadb==1.0.17
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
coverage==7.11.0
cycler==0.12.1
distro==1.9.0
Django==5.2.7
django-faker==0.2
django-filter==25.2
django-permissions-policy==4.28.0
django-redis==6.0.0
django-silk==5.4.3
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djongo==1.3.7
durationpy==0.10
et_xmlfile==2.0.0
factory_boy==3.3.3
fake-factory==9999.9.9
Faker==37.11.0
filelock==3.19.1
flake8==7.3.0
flatbuffers==25.2.10
fonttools==4.60.1
fsspec==2025.7.0
gast==0.6.0
google-auth==2.40.3
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gprof2dot==2025.4.14
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hiredis==3.3.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
isort==7.0.0
joblib==1.5.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.11.3
kiwisolver==1.4.9
kubernetes==33.1.0
legacy-cgi==2.6.3
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.3
mmh3==5.2.0
mpmath==1.3.0
mypy==1.18.2
mypy_extensions==1.1.0
namex==0.1.0
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.22.1
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.2
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
posthog==5.4.0
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycodestyle==2.14.0
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==3.11.4
pyparsing==3.2.5
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.4.0
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.4
tqdm==4.67.1
typer==0.16.0
types-PyYAML==6.0.12.20250915
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
zipp==3.23.0
